move shell_profile to compat/utils.rb & deprecate
